# obser-analystics/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a todo list management application",
        "title": "Todo Application",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/api",
    "paths": {
...
        "/hop-detail": {
            "get": {
                "description": "Get Hop Detail By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hop"
                ],
                "summary": "Get Hop Detail By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hop Id",
                        "name": "hop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit",
                        "name": "unit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HopDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/long-path": {
            "get": {
                "description": "Get Long Path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "path"
                ],
                "summary": "Get Long Path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Threshold",
                        "name": "threshold",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Path"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/path-detail": {
            "get": {
                "description": "Get Path Detail By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "path"
                ],
                "summary": "Get Path Detail By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Path Id",
                        "name": "path_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit",
                        "name": "unit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PathDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/paths": {
            "get": {
                "description": "Get All Path From Hop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "path"
                ],
                "summary": "Get All Path From Hop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Caller Service",
                        "name": "caller_service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Caller Operation",
                        "name": "caller_operation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Called Service",
                        "name": "called_service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Called Operation",
                        "name": "called_operation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Path"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Hop": {
            "type": "object",
            "properties": {
                "calledOperationName": {
                    "type": "string"
                },
                "calledServiceName": {
                    "type": "string"
                },
                "callerOperationName": {
                    "type": "string"
                },
                "callerServiceName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pathId": {
                    "type": "integer"
                }
            }
        },
        "model.HopDetail": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "distribution": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "errorCount": {
                    "type": "integer"
                },
                "errorDist": {
                    "description": "err code with count",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "errorRate": {
                    "type": "number"
                },
                "frequency": {
                    "type": "number"
                },
                "hopInfo": {
                    "$ref": "#/definitions/model.Hop"
                },
                "latency": {
                    "description": "max min p50 p99",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "model.Path": {
            "type": "object"
        },
        "model.PathDetail": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "distribution": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "errorCount": {
                    "type": "integer"
                },
                "errorDist": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "errorRate": {
                    "type": "number"
                },
                "frequency": {
                    "type": "number"
                },
                "pathInfo": {
                    "$ref": "#/definitions/model.Path"
                }
            }
        },
...
    }
}